+++ imod/formats/rasterio.py
@@ -7,7 +7,6 @@ Currently only :func:`imod.rasterio.write` is implemented.
 """
 
 import pathlib
-import warnings
 from typing import Dict
 
 import numpy as np
@@ -169,14 +168,6 @@ def _read(path, headersize, nrow, ncol, nodata, dtype):
         return array_io.reading._to_nan(a, nodata)
     else:
         return a
-
-
-def read(path, use_cftime=False, pattern=None):
-    warnings.warn(
-        "The imod.rasterio.read() function is deprecated. Instead use imod.rasterio.open().",
-        FutureWarning,
-    )
-    return open(path, use_cftime=False, pattern=None).load()
 
 
 def open(path, use_cftime=False, pattern=None):