+++ imod/mf6/__init__.py
@@ -52,7 +52,7 @@ from imod.mf6.riv import River
 from imod.mf6.simulation import Modflow6Simulation
 from imod.mf6.src import MassSourceLoading
 from imod.mf6.ssm import SourceSinkMixing
-from imod.mf6.sto import SpecificStorage, Storage, StorageCoefficient
+from imod.mf6.sto import SpecificStorage, StorageCoefficient
 from imod.mf6.timedis import TimeDiscretization
 from imod.mf6.uzf import UnsaturatedZoneFlow
 from imod.mf6.validation_settings import ValidationSettings